+++ test/test_helper.exs
@@ -1,2 +1,7 @@
+Supervisor.start_link([{EctoResource.TestRepo, []}],
+  strategy: :one_for_one,
+  name: EctoResource.Supervisor
+)
+
 ExUnit.start()
 Ecto.Adapters.SQL.Sandbox.mode(EctoResource.TestRepo, :manual)